+++ src/fr/cubiccl/generator3/game/object/global/GlobalBlock.java
@@ -5,12 +5,12 @@ import fr.cubiccl.generator3.util.Settings.Version;
 
 public class GlobalBlock extends GlobalObject
 {
-	public GlobalBlock(String id, double order)
+	public GlobalBlock(String id, int order)
 	{
 		super("block." + id, order);
 	}
 
-	public GlobalBlock(String id, double order, Version introduced, Version removed)
+	public GlobalBlock(String id, int order, Version introduced, Version removed)
 	{
 		super("block." + id, order, introduced, removed);
 	}
