+++ src/fr/cubiccl/generator3/game/object/global/GlobalEntity.java
@@ -5,12 +5,12 @@ import fr.cubiccl.generator3.util.Settings.Version;
 
 public class GlobalEntity extends GlobalObject
 {
-	public GlobalEntity(String id, double order)
+	public GlobalEntity(String id, int order)
 	{
 		super("entity." + id, order);
 	}
 
-	public GlobalEntity(String id, double order, Version introduced, Version removed)
+	public GlobalEntity(String id, int order, Version introduced, Version removed)
 	{
 		super("entity." + id, order, introduced, removed);
 	}
