+++ src/fr/cubiccl/generator3/game/object/global/GlobalNBTTag.java
@@ -5,12 +5,12 @@ import fr.cubiccl.generator3.util.Settings.Version;
 
 public class GlobalNBTTag extends GlobalObject
 {
-	public GlobalNBTTag(String id, double order)
+	public GlobalNBTTag(String id, int order)
 	{
 		super("nbttag." + id, order);
 	}
 
-	public GlobalNBTTag(String id, double order, Version introduced, Version removed)
+	public GlobalNBTTag(String id, int order, Version introduced, Version removed)
 	{
 		super("nbttag." + id, order, introduced, removed);
 	}
